convert dns server to plugin
